Fixed Building on windows
Since none of the dev team has previously been building on windows, there were a few errors which have now been corrected. This invovled renaming a directory and removing all unicode symbols. Additionally, this fixes the programs tendancy to always add .txt to the end of a file. It now only appends it if it doesn't see the .txt suffix