Fixed appending .txt logic
Previously, the program had a nasty habit of appending .txt to the end of every file, even if it already ended with .txt. I fixed this by adding some logic code to control it. Additionally, I fixed some line ending compatiblity that plagued our ending logic files and added a makefile for greater compatibility with windows.